pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-spectro-agent"
version = "0.1.0"
description = "LLM Spectro Agent for astronomical spectrum analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Author"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "asyncio>=4.0.0",
    "attrs>=25.3.0",
    "Authlib>=1.6.5",
    "beautifulsoup4>=4.13.5",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "colorlog>=6.10.1",
    "contourpy>=1.3.3",
    "cyclopts>=4.1.0",
    "dashscope>=1.24.5",
    "dataclasses-json>=0.6.7",
    "distro>=1.9.0",
    "einops>=0.8.1",
    "fastmcp>=2.13.0.2",
    "filetype>=1.2.0",
    "fonttools>=4.59.2",
    "fsspec>=2025.12.0",
    "ftfy>=6.3.1",
    "google-ai-generativelanguage>=0.7.0",
    "google-auth>=2.43.0",
    "greenlet>=3.2.4",
    "grpcio>=1.75.1",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "huggingface_hub>=1.2.3",
    "idna>=3.10",
    "Jinja2>=3.1.6",
    "jiter>=0.11.0",
    "joblib>=1.5.2",
    "json5>=0.12.1",
    "jsonpatch>=1.33",
    "jsonschema>=4.25.1",
    "jupyter>=1.1.1",
    "jupyter_client>=8.6.3",
    "jupyter_core>=5.8.1",
    "jupyterlab>=4.4.7",
    "langchain>=1.2.0",
    "langchain-community>=0.3.29",
    "langchain-core>=1.2.2",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.1.4",
    "langgraph>=1.0.3",
    "langsmith>=0.4.25",
    "lark>=1.2.2",
    "lightgbm>=4.6.0",
    "lxml>=6.0.2",
    "markdown-it-py>=4.0.0",
    "MarkupSafe>=3.0.2",
    "marshmallow>=3.26.1",
    "matplotlib>=3.10.6",
    "mcp>=1.19.0",
    "mistune>=3.1.4",
    "modelscope>=1.33.0",
    "multidict>=6.6.4",
    "mypy_extensions>=1.1.0",
    "nbclient>=0.10.2",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "nest-asyncio>=1.6.0",
    "networkx>=3.5",
    "notebook>=7.4.5",
    "numpy>=2.2.6",
    "openai>=2.8.0",
    "opencv-contrib-python>=4.10.0.84",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "orjson>=3.11.3",
    "packaging>=24.2",
    "pandas>=2.3.2",
    "pathvalidate>=3.3.1",
    "pillow>=11.3.0",
    "prompt_toolkit>=3.0.52",
    "propcache>=0.3.2",
    "protobuf>=6.32.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pydantic_core>=2.33.2",
    "pydantic-settings>=2.10.1",
    "Pygments>=2.19.2",
    "pyparsing>=3.2.3",
    "pypdfium2>=5.2.0",
    "pytesseract>=0.3.13",
    "python-bidi>=0.6.7",
    "python-dateutil>=2.9.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "PyYAML>=6.0.2",
    "referencing>=0.36.2",
    "regex>=2025.9.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "rpds-py>=0.27.1",
    "rsa>=4.9.1",
    "ruamel.yaml>=0.18.16",
    "safetensors>=0.7.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "sentencepiece>=0.2.1",
    "setuptools>=80.9.0",
    "shapely>=2.1.2",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "SQLAlchemy>=2.0.43",
    "stack-data>=0.6.3",
    "starlette>=0.48.0",
    "tenacity>=9.1.2",
    "terminado>=0.18.1",
    "threadpoolctl>=3.5.0",
    "tiktoken>=0.11.0",
    "tinycss2>=1.4.0",
    "tokenizers>=0.22.1",
    "tornado>=6.5.2",
    "tqdm>=4.67.1",
    "traitlets>=5.14.3",
    "typing_extensions>=4.15.0",
    "typing-inspect>=0.9.0",
    "tzdata>=2025.2",
    "ujson>=5.11.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.37.0",
    "wcwidth>=0.2.13",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
    "yarl>=1.20.1",
    "zstandard>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "debugpy>=1.8.16",
    "ipython>=8.12.3",
    "ipykernel>=6.30.1",
    "notebook>=7.4.5",
    "jupyterlab>=4.4.7",
]

[project.urls]
Homepage = "https://github.com/yourusername/llm-spectro-agent"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
